Fix problem with small font sizes in xterm
In newer version, the names of some resources seem to have changed.
Instead of a single resource for the font size, there are now multiple
ones.

All have the same name prefix followed by a digit, except the first
resource which represents the default font size.

The other resources specify the font sizes which can be selected from
the xterm menu when selection "unreadable" to "enourmous" as the font
size.